+++ jest.config.js
@@ -1,7 +1,7 @@
 module.exports = {
-  testEnvironment: "jsdom",
+  testEnvironment: "node",
   clearMocks: true,
-  collectCoverage: true,
+  collectCoverage: false,
   coverageDirectory: "coverage",
   coverageProvider: "v8",
   testMatch: ["**/__tests__/**/*.[jt]s?(x)", "**/?(*.)+(spec|test).[tj]s?(x)"],